+++ src/main/java/trend_setter/turtlerun/content/entity/DescriptionFile.java
@@ -1,9 +1,7 @@
-package trend_setter.turtlerun.global.content.entity;
+package trend_setter.turtlerun.content.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -15,25 +13,18 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "video_files")
-public class VideoFile {
+@Table(name = "description_files")
+public class DescriptionFile {
 
-    @Id @Column(name = "video_file_id")
+    @Id @Column(name = "description_file_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Enumerated(EnumType.STRING)
-    private VideoResolution videoResolution;
-
     private String fileName;
     private String filePath;
-    private int duration;
 
-    protected VideoFile(VideoResolution videoResolution, String fileName, String filePath,
-        int duration) {
-        this.videoResolution = videoResolution;
+    public DescriptionFile(String fileName, String filePath) {
         this.fileName = fileName;
         this.filePath = filePath;
-        this.duration = duration;
     }
 }
