+++ src/main/java/trend_setter/turtlerun/content/entity/VideoFile.java
@@ -1,7 +1,9 @@
-package trend_setter.turtlerun.global.content.entity;
+package trend_setter.turtlerun.content.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -13,18 +15,25 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "description_files")
-public class DescriptionFile {
+@Table(name = "video_files")
+public class VideoFile {
 
-    @Id @Column(name = "description_file_id")
+    @Id @Column(name = "video_file_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+    @Enumerated(EnumType.STRING)
+    private VideoResolution videoResolution;
+
     private String fileName;
     private String filePath;
+    private int duration;
 
-    public DescriptionFile(String fileName, String filePath) {
+    protected VideoFile(VideoResolution videoResolution, String fileName, String filePath,
+        int duration) {
+        this.videoResolution = videoResolution;
         this.fileName = fileName;
         this.filePath = filePath;
+        this.duration = duration;
     }
 }
